Make INSTALLLEVEL and REBOOT properties configurable
As with other properties in the wix template, it can
be useful for users of this library to override both
the reboot behavior (e.g. in cases they want to allow
reboots), as well as the install level (e.g. if they
want to exclude install level 2 features or include
install level 3 features without customers having to
use ADDLOCAL or other manual mechanisms).

This change makes both of those properties configurable,
defaulting them to their existing behavior.

It also cleans up a leftover comment (there were two sets
of descriptions for INSTALLPATH; this change removes the
older, misplaced version.